feat: generate post json
