+++ packages/docs/src/entry.ssr.tsx
@@ -16,8 +16,35 @@ import {
 } from "@builder.io/qwik/server";
 import { manifest } from "@qwik-client-manifest";
 import Root from "./root";
+import { DocumentHeadProps } from "@builder.io/qwik-city";
+import { asyncMap } from "./util";
+import fs from "fs";
 
-export default function (opts: RenderToStreamOptions) {
+export default async function (opts: RenderToStreamOptions) {
+  const modules = await import.meta.glob(
+    "/src/routes/blog/contents/**/**/index.mdx"
+  );
+  const posts = (
+    await asyncMap(Object.keys(modules), async (path) => {
+      const data = (await modules[path]()) as DocumentHeadProps;
+      return {
+        title: data.head.title || "",
+        description:
+          data.head.meta.find((m) => m.name === "description")?.content || "",
+        permalink: path
+          .replace(/^\/src\/routes/, "")
+          .replace(/\/index.mdx$/, "/"),
+        date: data.head.frontmatter.date,
+        tags: data.head.frontmatter.tags,
+        published: data.head.frontmatter.published,
+      };
+    })
+  ).sort((a, b) => {
+    const dateA = new Date(a.date).getTime();
+    const dateB = new Date(b.date).getTime();
+    return dateB - dateA;
+  });
+  fs.writeFileSync("./src/routes/blog/posts.json", JSON.stringify(posts));
   return renderToStream(<Root />, {
     manifest,
     ...opts,
